% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.BA.r
\name{module.BA}
\alias{module.BA}
\title{Predicts and projects stand basal area based on stand-level information.}
\usage{
module.BA(N0 = NA, HDOM0 = NA, projection = FALSE, BA0 = NA, N1 = NA,
  HDOM1 = NA)
}
\arguments{
\item{N0}{Numeric value of number of trees per hectare at age 0 (or initial age).}

\item{HDOM0}{Numeric value of Dominant Height (m) at age 0.}

\item{projection}{If TRUE then model projection from provided BA0 is executed for a 1 year increment. Default: FALSE.}

\item{BA0}{Numeric value of Basal Area (m2/ha) at age 0 (required for model projection).}

\item{N1}{Numeric value of number of trees per hectare at age 1 for projection.}

\item{HDOM1}{Numeric value of Dominant Height (m) at age 1 for projection.}
}
\value{
A list containing:

\code{- BA0} Predicted Basal Area at age 0 (m2/ha).

\code{- BA1} Projected Basal Area at age 1 (m2/ha).

}
\description{
\code{module.BA} Predicts and projects stand basal area based on number of trees per hectare and
dominant height at the stand-level using the equation: ln(BA) = c1 + c2 x ln(N) + c3 x ln(HDOM)
}
\examples{
# Example 1 - Predicting BA
module.BA(N0=1200, HDOM0=17.7)$BA0

# Example 2 - Projecting BA
module.BA(N0=1200, HDOM0=17.7, projection=TRUE, BA0=24, N1=1182, HDOM1=18.9)$BA1

# Example 3 - Varifying projection with prediction
sim.test <- module.input(TYPE='PLOT', BA0=17.63402, HDOM=14, AGE0=17, N0=1200, AGEF=28)
module.simulation(stand=sim.test)$sim.stand                   # All projections
module.BA(N0=1200, HDOM0=14, projection=FALSE)$BA0            # Prediction Age 17
module.BA(N0=1185.363, HDOM0=14.75542, projection=FALSE)$BA0  # Prediction Age 18
module.BA(N0=1019.661, HDOM0=21.26901, projection=FALSE)$BA0  # Prediction Age 28
}
\references{
Gonzalez-Benecke et al. (2012) Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests 3(4), 1104-1132; doi:10.3390/f3041104
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
