% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.input.r
\name{module.input}
\alias{module.input}
\title{Module of input tree- or stand-level data to prepare it for further simulations.}
\usage{
module.input(TYPE = "PLOT", TREEDATA = NA, AREA = NA, SI = NA,
  HDOM0 = NA, AGE0 = NA, BA0 = NA, N0 = NA, AGEF = 50,
  THINNING = "FALSE", AGET = NA, BAR = NA, t = 5, d = 15,
  method = 2)
}
\arguments{
\item{TYPE}{Character for type of input data. PLOT: stand-level data information, TREE: tree-level information. Default is PLOT}

\item{TREEDATA}{Data frame with tree-level information with columns: PLOTID, TREEID, DBH, HT (these should be identical names).}

\item{AREA}{Numeric value of size of the inventory plot (m2). Required for TYPE='TREE'.}

\item{SI}{Numeric value of Site Index (m) (Dominant Height of the plot at age 50 years).}

\item{HDOM0}{Numeric value of Dominant Height (m) at initial age (or age 0).}

\item{AGE0}{Numeric value of initial stand age or age 0 (years).}

\item{BA0}{Numeric value of Basal Area (m2/ha) at age 0 (required for model projection).}

\item{N0}{Numeric value of number of trees per hectare at age 0.}

\item{AGEF}{Numeric value of final stand age (in years) of simulation. Default is 50.}

\item{THINNING}{If TRUE then a thinning is implemented according to AGET and BAR. Default is FALSE.}

\item{AGET}{Numeric value of stand age (in years) where thinning is planned.}

\item{BAR}{Numeric value of Relative Basal Area (\%, 0-1) to be removed when thinning at age AGET.}

\item{t}{Numeric value top stem diameter outside bark for merchantability limit (cm).}

\item{d}{Numeric value of a DBH threshold limit for merchantable trees (cm).}

\item{method}{Numeric value that identifies the method to estimate missing heights from TYPE='TREE'.
1: parametrized DBH-height model that requires DBH, BA and AGE, 2: fits a simple DBH-height model from 
available measurements using the equation: ln(Ht) = b0 + b1/DBH. Default method=2.}
}
\value{
A list containing the following: 
\itemize{   
\item \code{SI}       Site Index (m).
\item \code{AGE0}     Initial stand age or age 0 (years).
\item \code{HDOM0}    Dominant Height (m) at initial age (or age 0). 
\item \code{BA0}      Basal Area (m2/ha) at age 0.
\item \code{N0}       Number of trees per hectare at age 0. 
\item \code{QD}       Mean quadratic diameter (cm) at age 0.
\item \code{SDIR0}    Relative stand density index (\%) at age 0.
\item \code{VOL_OB0}  Total stand-level volume outside bark (m3/ha) at age 0.
\item \code{VOL_IB0}  Total stand-level volume inside bark (m3/ha) at age 0.
\item \code{VOLm_OB0} Merchantable stand-level volume outside bark (m3/ha) at age 0.
\item \code{VOLm_IB0} Merchantable stand-level volume inside bark (m3/ha) at age 0.
\item \code{AGEF}     Final stand age (in years) of simulation.
\item \code{THINNING} Logical that indicates if thinning is implemented according to AGET and BAR.
\item \code{AGET}     Stand age (in years) where thinning is planned.
\item \code{BAR}      Relative Basal Area (\%, 0-1) to be removed when thinning at age AGET.
\item \code{t}        Top stem diameter outside bark for merchantability limit (cm).
\item \code{d}        DBH threshold limit for merchantable trees (cm).
\item \code{method}   Selection of the method to estimate missing heights from TYPE='TREE'.
}
}
\description{
\code{module.input} Prepares tree- or stand-level data from a single plot, checks and completes missing values, and 
calculates several stand-level parameters including total volume. It also reads required information for further 
simulations including simulation age and details of future thinning. Some information is only traspassed to other modules.
Note that form tree-level data individual tree (complete or incomplete) information is required.
}
\examples{
# Example 1 - Input stand-level data
module.input(TYPE='PLOT', BA0=17.3, SI=30, AGE0=17, N0=1200, AGEF=18)
module.input(TYPE='PLOT', BA0=17.3, HDOM0=16, AGE0=17, N0=1200, AGEF=18)
module.input(TYPE='PLOT', HDOM0=16, AGE0=17, N0=1200, AGEF=18)  # BA obtained by prediction

# Example 2 - Input with individual tree data
module.input(TYPE='TREE', TREEDATA=treedata, AREA=500, AGE0=23, AGEF=32)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
