% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.N.r
\name{module.N}
\alias{module.N}
\title{Predicts the number of trees (mortality) at next age based on stand-level information.}
\usage{
module.N(N0 = NA, HDOM0 = NA, SDIR0 = NA, AGE0 = NA, AGE1 = NA)
}
\arguments{
\item{N0}{Numeric value of number of trees per hectare at age 0.}

\item{HDOM0}{Numeric value of Dominant Height (m) at age 0.}

\item{SDIR0}{Numeric value of Relative Stand Density Index for Longleaf (\%) at age 0.}

\item{AGE0}{Numeric value of intial stand age or age 0 (years).}

\item{AGE1}{Numeric value of final stand age or age 1 for prediction (years).}
}
\value{
A value with the number of trees per hectare at age 1 (N1).
}
\description{
\code{module.N} Estimates the number of trees (mortality) from initial age (AGE0) to next age (AGE1) based on stand level 
information based on the equation:
N1 = N0 x exp((c1x(HDOM0/100)+c2xSDIR) x ((AGE1^c3)-(AGE0^c3)))
}
\examples{
module.N(N0=2500, HDOM0=14, SDIR0=45, AGE0=24, AGE1=25)$N1
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
