% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.VOL.r
\name{module.VOL}
\alias{module.VOL}
\title{Calculates the total stand-level volume based on stand-level information.}
\usage{
module.VOL(N = NA, BA = NA, AGE = NA, SI = NA)
}
\arguments{
\item{N}{Numeric value of number of trees per hectare.}

\item{BA}{Numeric value of Basal Area of the plot (m2/ha).}

\item{AGE}{Numeric value of Age of the plot (in years).}

\item{SI}{Numeric value of Site Index (m) (Dominant Height of the plot at age 50 years).}
}
\value{
A list containing the parameters:
\itemize{ 
\item \code{VOL_OB} Total stand-level volume outside bark (m3/ha).
\item \code{VOL_IB} Total stand-level volume inside bark (m3/ha).
}
}
\description{
\code{module.VOL} Calculates the total stand-level volume inside and outside bark for a plot 
based on the equation: ln(VOL) = d1 + d2xln(N) + d3xln(BA) + d4xln(BA/AGE) + d5xln(SI)
}
\examples{
module.VOL(N=2500, BA=30, AGE=25, SI=15)
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
