% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.VOLm.r
\name{module.VOLm}
\alias{module.VOLm}
\title{Calculates the stand-level merchantable volume based on stand-level information.}
\usage{
module.VOLm(N = NA, QD = NA, t = NA, d = NA, VOL_OB = NA,
  VOL_IB = NA)
}
\arguments{
\item{N}{Numeric value of number of trees per hectare.}

\item{QD}{Numeric value of mean Quadratic Diameter (cm).}

\item{t}{Numeric value top stem diameter outside bark for merchantability limit (cm).}

\item{d}{Numeric value of a DBH threshold limit for merchantable trees (cm).}

\item{VOL_OB}{Total stand-level volume outside bark (m3/ha).}

\item{VOL_IB}{Total stand-level volume inside bark (m3/ha).}
}
\value{
A list containing the parameters:

\code{- VOLm_OB} Merchantable stand-level volume outside bark (m3/ha).

\code{- VOLm_IB} Merchantable stand-level volume inside bark (m3/ha).

}
\description{
\code{module.VOLm} Calculates the stand-level merchantable volume inside and outside bark fro a plot
based on the equation:  VOLm = VOL x exp(m1 x ((t/QD)^m2) + m3 x (N^m4) x (d/QD)^m5)
}
\examples{
VOL_OB<-module.VOL(N=2500, BA=30, AGE=25, SI=15)$VOL_OB
VOL_IB<-module.VOL(N=2500, BA=30, AGE=25, SI=15)$VOL_IB
QD<-stand.STAND(BA=30, N=2500)$QD
module.VOLm(N=2500, QD=QD, t=3, d=5, VOL_OB=VOL_OB, VOL_IB=VOL_IB)
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\seealso{
\code{\link{module.VOL}}
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
