% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.SDI.r
\name{stand.SDI}
\alias{stand.SDI}
\title{Calculates relative stand density index for Longleaf Pine (\emph{Pinus palustris})}
\usage{
stand.SDI(N = NA, QD = NA)
}
\arguments{
\item{N}{Numeric value of the number of trees per hectare.}

\item{QD}{Numeric value of mean quadratic diameter (cm).}
}
\value{
A value for the relative stand density index (RSDI, \%).
}
\description{
\code{stand.SDI} Calculates the relative stand density index (RSDI, \%) for Longleaf in a single plot using the
expression: RSDI = 100xNx(QD/25.4)^1.605/SDImax. For Longleaf pine (\emph{Pinus palustris}) SDImax is 1200 trees/ha. 
At least two parameters are required to complete the missing.
}
\examples{
stand.SDI(QD=18, N=1660)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
