% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.SITE.r
\name{stand.SITE}
\alias{stand.SITE}
\title{Calculates dominant height, site index and stand age for a plot.}
\usage{
stand.SITE(HDOM = NA, SI = NA, AGE = NA)
}
\arguments{
\item{HDOM}{Numeric value of mean Dominant Height (m).}

\item{SI}{Numeric value of Site Index (m) (Dominant Height of the plot at age 50 years).}

\item{AGE}{Numeric value of stand Age (years).}
}
\value{
A list containing the parameters: 
\itemize{
\item \code{HDOM} Dominant Height (m).
\item \code{SI}   Site Index (m).
\item \code{AGE}  Age (years).
}
}
\description{
\code{stand.SITE} Calculates dominant height, site index (at reference age 50 year) or stand age for a plot 
based on the equation following dominant height equation: 
HDOM = SI x ((1-exp(a1 x Age)) / (1-exp(a1 x 50)))^a2
The definition of HDOM is based on the top 25th percentile of the tree heights. 
At least two parameters are required to calculate the thrid missing parameter.
}
\examples{
stand.SITE(SI=30,AGE=40)$HDOM     # Missing HDOM
stand.SITE(HDOM=27.5,AGE=40)$SI   # Missing SI
stand.SITE(HDOM=27.5,SI=33.3)$AGE # Missing AGE
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
