% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.STAND.r
\name{stand.STAND}
\alias{stand.STAND}
\title{Calculates the mean quadratic diameter, basal area or number of trees.}
\usage{
stand.STAND(BA = NA, N = NA, QD = NA)
}
\arguments{
\item{BA}{Numeric value of basal area of the plot (m2/ha).}

\item{N}{Numeric value of number of trees per hectare.}

\item{QD}{Numeric value of mean quadratic diameter (cm).}
}
\value{
A list containing the parameters: 
\itemize{
\item \code{BA} Basal Area (m2/ha).
\item \code{N}  Number of trees per hectare.
\item \code{QD} Mean Quadratic Diameter (cm).
}
}
\description{
\code{stand.STAND} Calculates the mean quadratic diameter, basal area or number of trees based on stand-level information 
of the plot. At least two parameters are required to calculate the third missing parameter.
}
\examples{
stand.STAND(QD=18, N=1660)$BA   # Missing BA 
stand.STAND(BA=42, QD=18)$N     # Missing N 
stand.STAND(BA=42, N=1660)$QD   # Missing QD
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
