% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.HDOM.r
\name{tree.HDOM}
\alias{tree.HDOM}
\title{Calculate dominant height for a plot based on tree-level data.}
\usage{
tree.HDOM(HT)
}
\arguments{
\item{HT}{Vector of tree heights (m) for a given plot (must be complete).}
}
\value{
A value with Dominant Height (HDOM, m) for the plot.
}
\description{
\code{tree.HDOM} Calculate the dominant height for a plot based on tree-level data based on the top 25th percentile
of the tree heights. The provided vector of heights should be complete without missing data.
}
\examples{
HT <- treedata$HT
tree.HDOM(HT=HT)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
