% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.HDOM.m.r
\name{tree.HDOM.m}
\alias{tree.HDOM.m}
\title{Calculate dominant height for a plot based on tree-level data.}
\usage{
tree.HDOM.m(HT = NA, DBH = NA, AREA = NA)
}
\arguments{
\item{HT}{Vector of tree heights (m) for a given plot (must be complete).}

\item{DBH}{Vector of diameter at breast height (DBH, cm). Must be complete and have the same size and order of HT.}

\item{AREA}{Numeric value of area of the inventory plot (m2).}
}
\value{
A value with Dominant Height (HDOM, m) for the plot.
}
\description{
\code{tree.HDOM.m} Calculate the dominant height for a plot based on tree-level data based on the top 25th percentile
of the tree heights. The provided vector of heights should be complete without missing data.
}
\examples{
HT <- tree.HT(DBH=treedata$DBH, HT=treedata$HT, method=2)$HTFIN
DBH <- treedata$DBH
AREA <- 301
tree.HDOM.m(HT=HT, DBH=DBH, AREA=AREA)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
